TSK-1812: Added details in task-history for claim/cancel
Detect changes and pass them in the specific event in TaskServiceImpl class. For saving the old state of the specified task, I had to use copy() since getTask returns the same object reference, thus no changes can be detected. 
In the specified events details are accepted as parameters.
Extended Tests for Claim and Cancel Events to check correct details.